Add pagination support for listing container instances
We've been seeing intermittent 504 errors on one of our services
deployed via ECS. After looking into the Lambda logs, I noticed
that there were a bunch of errors in the form of:

```
i-09432453c14d92f6e" not found in the cluster "production": errorString null
```

This led me to the `getContainerInstance` function, which I noticed had
no pagination. The max returned per page is 100, so any time a container
instance was terminated and didn't happen to be in the first page, the
container would not transition to draining and the tasks were killed
abruptly.

The fix is to use the `ListContainerInstancesPages` function to ensure
we fetch all container instance: https://docs.aws.amazon.com/sdk-for-go/api/service/ecs/#ECS.ListContainerInstancesPages